+++ types/collection/descriptive.ts
@@ -1,7 +1,9 @@
-import type { Collection, Distinct, Group } from '../abstract/collection';
+import type { Collection, Distinct, Group, Single } from '../abstract/collection';
+import type { PrimitiveProperty } from '../abstract/property';
 import type { RefId } from '../abstract/reference';
+import type { LangGroup } from '../abstract/util';
 import type { LangCode } from '../enum/generic';
-import type { LangGroup, MediaGroup, WeblinksGroup } from './generic';
+import type { D3Format, ImageFormat } from '../enum/util';
 import type { RegistryGroup, StructureGroup } from './registry';
 
 export type DiscoveryGroup = Group< {
@@ -10,6 +12,44 @@ export type DiscoveryGroup = Group< {
   country?: Distinct< string | string[] >;
   institute?: Distinct< string | string[] >;
   references?: RefId[];
+} >;
+
+export type MediaGroup = Group< {
+  images?: Distinct< {
+    url: string;
+    format?: ImageFormat;
+    credits: string;
+    license: string;
+    author?: string;
+    source?: string;
+    width?: number;
+    height?: number;
+  }[] >;
+
+  spectrum?: Group< {
+    absorption?: Single< PrimitiveProperty< string > >;
+    emission?: Single< PrimitiveProperty< string > >;
+    uv?: Single< PrimitiveProperty< string > >;
+    xray?: Single< PrimitiveProperty< string > >;
+  } >;
+
+  structure3D?: Distinct< {
+    data: string;
+    format: D3Format;
+    url?: string;
+  }[] >;
+} >;
+
+export type WeblinksGroup = Group< {
+  links?: Distinct< {
+    url: string;
+    text?: string;
+    description?: string;
+    archiveUrl?: string;
+    accessed?: string;
+    language?: string;
+  }[] >;
+  wiki?: LangGroup;
 } >;
 
 export type DescriptiveCollection = Collection< {
