+++ types/entity/element.ts
@@ -1,14 +1,15 @@
 import type { Expand } from 'devtypes/types/util';
 import type { Collection, Distinct } from '../abstract/collection';
+import type { MetaData } from '../abstract/util';
 import type { AbundanceCollection } from '../collection/abundance';
 import type { AtomicsCollection } from '../collection/atomics';
 import type { ChemistryCollection } from '../collection/chemistry';
 import type { DescriptiveCollection } from '../collection/descriptive';
-import type { GenericCollection, MetaData } from '../collection/generic';
+import type { GenericCollection } from '../collection/generic';
 import type { PhysicsCollection } from '../collection/physics';
 import type { SafetyCollection } from '../collection/safety';
-import type { Phase, PTColumn, PTPeriod } from '../enum/generic';
 import type { ElementBlock, ElementGroup, ElementProperty, ElementSet, ElementSymbol } from '../enum/element';
+import type { Phase, PTColumn, PTPeriod } from '../enum/generic';
 
 export type ElementClassification = Collection< {
   symbol: Distinct< string >;
